requirements.txt
gradio
rembg
Pillow
